+++ CHANGES.rst
@@ -364,6 +364,12 @@ Bug Fixes
   - Change construction of result of arithmetic and unit conversion to allow
     subclasses to require the presence of attribute like unit. [#2300]
 
+  - Scale uncertainties to correct units in arithmetic operations and unit
+    conversion. [#2393]
+
+  - Ensure uncertainty and mask members are copied in arithemetic and
+    convert_unit_to. [#2394]
+
 - ``astropy.stats``
 
 - ``astropy.table``